+++ NEWS.md
@@ -1,3 +1,6 @@
+## Bug Fixes
+- fixed retrieval of access token when multiple access tokens are stored
+
 # RKorAPClient 0.6.1
 
 ## Changes